Update Views
This is a big update to the Views pages.
- now uses standard 2 panel layout for adding/editing/viewing
Views (includes the sidebar)
- New menu added to all Views, containing [Edit, Clone, Delete]
options. Clone and Delete are their own modals. This modal is
extensible, so can be used for different types of indexes added
via extensions.
- sidebar now properly selects views and highlights them to
show current location + view.
- Renaming a view no longer creates a second view, just renames
it. The clone feature lets you copy/clone a View.
- cog icon in header replaced with a 3-dot thing.

N.B. the new menu uses React Bootstrap so it's very similar
but not identical to our current menu dropdowns. The latter
will be replaced in subsquent PRs.